feat : decrease point, create pointHistory, reward transaction

+++ ad-hell/src/main/java/com/adhd/ad_hell/domain/user/command/dto/request/UserPointRequest.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class UserPointRequest {
-  private Long point;
+  private Integer point;
 }
